made everything work for new data structure.
I changed the nonevoked sessions in a way that it also gives the data which is not averaged over trials. Had to change a lot also in nonevoked_syn_analysis in order to make it work. A problem was also that chalva has only one handposition, now the script should also work for this.

*fixed rank1 tests to pronation handposition