changes of all_data data structure.
in order to concatenate the structs I had to quarantee that
they all have the same fields when I don't want to have a test substruct.
I thought without the substruct I might have more elegant access to the
data, but I realized it does not work. matlab is ugly

the data for both handpositions together is not collected anymore.