changed the indexing for selection sessions of a monkey
before I had it in a vector with the indeces. This could
lead to problems when I removed sessions in the beginning.
To avoid this I store the indeces now in logical vectors
which all have equal length. By doing this I also had to change
a few other lines in the file, where amount of trials was
displayed or only special sessions of a monkey were selected