analysis of settings of the pcaica (runica method)
it was unstable with this extended 1 setting, so I 
removed and with this also got rid of an error message.

changed also the pcaica method as I tried to use the runica
builtin pca. But gives worse results (concerning stability
over sessions which appears for nmf) and so I used yuvals.
But still I have to check whether the algo of yuval is correct