match groups of vectors by corrcoef and dot product
first I used the dot product for matching and for matching score, than corrcoef for matching score as in the tresch paper and then I wondered why not for everything. Now I made it an option of the function, what method to choose. dot product gives better results.

I added a test function in method_validations to prove this