fix a bug introduced in 8490105634f6b04
here we decided to only use the data_raw (responses not averaged over trials). I did not remove all the code properly

also remove the target statistics

because I don't remember anymore what it really means and I think it also does not make sense if we only use the data_raw struct which is not averaged over trials.