merge in yifats changes of the run script for emgdat.
1.The program that generate the muscle fileds is called runscript4evoked. I am now working on my own version (in a folder which is copy of stephan folder). I made the following changes:
a in the routine: “stimulations_at” I added a line which skip filtering in case stim amplitude is empty.
b. in get_all_stimsess I am sorting the session folders before generating the data structure. This is done by a utilitiy sortdirsyp now locatde in ..\lib\ folder (the original utility sortdirs of that folder doesn’t work well). Some changes in the routine were required to match the output of sortdirs.
c. in responses.m I have a problem because in some of darma bhv files there are no stim_time field. This needs to be fixed!
d. the output file name is now called ‘allevoked....mat’ because it contains all stim amps and not only the 150 microamps
e. I am now putting all results in d:\stephan\resultsyp\ that is a contains a copy of the data from stephan + my added data
f. in responses.m I fixed a bug that happened rarely in cases were we had 3 stim files.