investigate the influence of the stimulations strength
we were regularly asked whether we are sure that we stimulate only
single neurons. Yifat thinks that we stimulate single neurons
or at least small groups. This is an important point because if we would
already stimulate groups of neurons the synergies we obsever could
arise from (linear) combinations of activity patterns of neurons.
To check this the influence of the stimulation amplitude is investigated.
-> The larger the amplitude the more neurons should be stimulated
-> if the activation pattern is a combination of singel neuron patterns, it should
   change when more neurons are activated.
-> therefore this function checks whether the stimulation amplitude influences
   the direction of the activation pattern (evidence that it is combined from single
    neuron activation patterns) or that it influences the only the amplitude
the idea is to use a svd to check the rank of the matrix which indicates the linear
dependency of its rows

 calling example
 stimamp_influence('~/projects/yifat_paper/results/data/', ...
                   '~/projects/yifat_paper/results/evoked_syns/',  {'vega', 'chalva'})