normalization revisited: response correlations
as many times before, the issue or normalization came up.
We discussed whether we should normalize the data from the
natural movements, if yes -> how and also whether to normalize the evoked
responses. Can we do this? Are the differences between the channels (some of them
are more active then others) correlated for natural movement data and evoked responses.
this file computes the correlation between the two and as a result showed
that means and stds are not correlated.