fix a bug for reading evoked_response files
the problem was that StimTime was sometimes a column instead of a row and the old code skipped these sessions. This caused that we missed ten sessions.

Fixing this bug gives us 10 sessions more. Because we have more sessions now this fits not anymore the size of the hand sorted struct so we had to manually sort it again for artifacts. This time I sorted less conservative and chose also sessions that might be slightly influenced by the artifact, if the show a nice response in all other channels..

I also add this sort files to the data directory because they might influence the results